requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn>=1.4.0
xgboost>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
PyYAML>=6.0
mlflow>=2.12.0
python-multipart>=0.0.9
pytest>=8.0.0